pyproject.toml
[tool.poetry]
name = "cvgen"
version = "0.4.0"
description = "A tool for easily creating a good-looking CV in PDF format from .yaml data, HTML & CSS."
authors = ["Vadim Galaktionov <vadim@galaktionov.nl>"]
include = ["cvgen/templates/*"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
weasyprint = "^47.0"
pyyaml = "^5.1"
click = "^7.0"
pydantic = "^0.25.0"
jinja2 = "^2.10"
email-validator = "^1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^3.0"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
mypy = "^0.701.0"

[tool.poetry.scripts]
cvgen = "cvgen.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
